Refactor sidebar component structure
